cleaner page's and security

+++ app/Http/Controllers/UserController.php
@@ -10,6 +10,8 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Redirect;
 use Spatie\Permission\Models\Role;
 use DB;
 use Hash;
@@ -20,32 +22,40 @@ class UserController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return Application|Factory|View|Response
+     * @return Application|Factory|View|\Illuminate\Http\RedirectResponse
      */
     public function index(Request $request)
     {
-        $data = User::orderBy('id','DESC')->paginate(5);;
-        return view('user.index',compact('data'))
-            ->with('i', ($request->input('page', 1) - 1) * 5);
+        $user = Auth::user();
+        if ($user->getRoleNames()[0] === 'Admin') {
+            $data = User::orderBy('id', 'DESC')->paginate(5);;
+            return view('user.index', compact('data'))
+                ->with('i', ($request->input('page', 1) - 1) * 5);
+        }
+        return Redirect::back()->with('danger', 'U heeft niet de juiste rechten om deze pagina te bezoeken');
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return Application|Factory|View|Response
+     * @return Application|Factory|View|\Illuminate\Http\RedirectResponse
      */
     public function create()
     {
-        $roles = Role::pluck('name','name')->all();
-        $groups = group::pluck('name','name')->all();
-        return view('user.create',compact('roles', 'groups'));
+        $user = Auth::user();
+        if ($user->getRoleNames()[0] === 'Admin') {
+            $roles = Role::pluck('name', 'name')->all();
+            $groups = group::pluck('name', 'name')->all();
+            return view('user.create', compact('roles', 'groups'));
+        }
+        return Redirect::back()->with('danger', 'U heeft niet de juiste rechten om deze pagina te bezoeken');
     }
 
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
@@ -65,110 +75,121 @@ class UserController extends Controller
         $user->assignRole($request->input('roles'));
 
         return redirect()->route('users.index')
-            ->with('success','User created successfully');
+            ->with('success', 'User created successfully');
     }
 
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return Application|Factory|View|Response
+     * @param int $id
+     * @return Application|Factory|View|\Illuminate\Http\RedirectResponse
      */
     public function show($id)
     {
         $user = User::find($id);
-        return view('user.show',compact('user'));
+        return view('user.show', compact('user'));
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return Application|Factory|View|Response
+     * @param int $id
+     * @return Application|Factory|View|\Illuminate\Http\RedirectResponse
      */
     public function edit($id)
     {
-        $user = User::find($id);
-        $roles = Role::pluck('name','name')->all();
-        $group = group::pluck('name','name')->all();
-        $userRole = $user->roles->pluck('name','name')->all();
-        $userGroup = $user->groups->pluck('name','name')->all();
+        $user = Auth::user();
+        if ($user->getRoleNames()[0] === 'Admin' || $user->id === (int)$id) {
+            $user = User::find($id);
+            $roles = Role::pluck('name', 'name')->all();
+            $group = group::pluck('name', 'name')->all();
+            $userRole = $user->roles->pluck('name', 'name')->all();
+            $userGroup = $user->groups->pluck('name', 'name')->all();
 
-        return view('user.edit',compact('user','roles','userRole', 'group' , 'userGroup'));
+            return view('user.edit', compact('user', 'roles', 'userRole', 'group', 'userGroup'));
+        }
+        return Redirect::back()->with('danger', 'U heeft niet de juiste rechten om deze pagina te bezoeken');
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request, $id)
     {
         $this->validate($request, [
             'name' => 'required',
-            'email' => 'required|email|unique:users,email,'.$id,
+            'email' => 'required|email|unique:users,email,' . $id,
             'password' => 'same:confirm-password',
             'roles' => 'required',
         ]);
 
         $input = $request->all();
 
-        if(!empty($input['password'])){
+        if (!empty($input['password'])) {
             $input['password'] = Hash::make($input['password']);
-        }else{
-            $input = Arr::except($input,array('password'));
+        } else {
+            $input = Arr::except($input, array('password'));
         }
 
         $user = User::find($id);
 
         $user->groups()->detach();
-        foreach ($input['groups'] as $group)
-        {
-            $group_id = DB::table('groups')->where('name',$group)->pluck('id');
+        foreach ($input['groups'] as $group) {
+            $group_id = DB::table('groups')->where('name', $group)->pluck('id');
 
             $user->groups()->attach($group_id);
 
         }
         $user->update($input);
-        DB::table('model_has_roles')->where('model_id',$id)->delete();
+        DB::table('model_has_roles')->where('model_id', $id)->delete();
 
         $user->assignRole($request->input('roles'));
 
         return redirect()->route('users.index')
-            ->with('success','User updated successfully');
+            ->with('success', 'User updated successfully');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy($id)
     {
-        User::find($id)->delete();
-        return redirect()->route('users.index')
-            ->with('success','User deleted successfully');
+        $user = Auth::user();
+        if ($user->getRoleNames()[0] === 'Admin') {
+            User::find($id)->delete();
+            return redirect()->route('users.index')
+                ->with('success', 'User deleted successfully');
+        }
+        return Redirect::back()->with('danger', 'U heeft niet de juiste rechten om deze pagina te bezoeken');
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return Application|Factory|View|Response
+     * @return Application|Factory|View|\Illuminate\Http\RedirectResponse
      */
     public function CreateList()
     {
-        $groups = group::pluck('name','name')->all();
-        return view('user.createList',compact('groups'));
+        $user = Auth::user();
+        if ($user->getRoleNames()[0] === 'Admin') {
+            $groups = group::pluck('name', 'name')->all();
+            return view('user.createList', compact('groups'));
+        }
+        return Redirect::back()->with('danger', 'U heeft niet de juiste rechten om deze pagina te bezoeken');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\RedirectResponse
      */
     public function StoreList(Request $request)
@@ -178,20 +199,18 @@ class UserController extends Controller
             'groups' => 'required'
         ]);
         $input = $request->all();
-        $students = str_replace( ' ', '', explode(',', $input['name']));
-        foreach ($students as $student)
-        {
+        $students = str_replace(' ', '', explode(',', $input['name']));
+        foreach ($students as $student) {
             $user = User::create([
                 'name' => $student,
                 'email' => $student . "@mydavinci.nl",
                 'password' => Hash::make('Welkom123'),
-                ]);
+            ]);
 
-            $user->assignRole($request->input('roles'));
+            $user->assignRole('User');
 
-            foreach ($input['groups'] as $group)
-            {
-                $group_id = DB::table('groups')->where('name',$group)->pluck('id');
+            foreach ($input['groups'] as $group) {
+                $group_id = DB::table('groups')->where('name', $group)->pluck('id');
 
                 $user->groups()->attach($group_id);
 
@@ -200,7 +219,7 @@ class UserController extends Controller
         }
 
         return redirect()->route('users.index')
-            ->with('success','User created successfully');
+            ->with('success', 'User created successfully');
 
     }
 }
